Switch back to the old partition cost calculation.
The old calculation (from the C++ version) is the sum of the costs for
each of the two bounding boxes produced from a partition split, where
each cost is the box surface area times the number of primitives in it.
Surface area is used to approximate the probability that a ray will
intersect a box.

The rust code had attempted to refine the calculation by excluding box
faces that overlapped each other, but ignored the fact that adjacent
faces won't overlap completely for different-sized boxes. It also mis-
counted the primitives in each box, leading to too-high costs that made
the algorithm produce no splits.